test(agents): split computer tool coverage
