+++ src/agents/tools/computer-tool.test.ts
@@ -1,129 +1,26 @@
-/**
- * computer tool tests.
- *
- * Cover the computer.act wire mapping, frame binding, and enablement behavior.
- * Node selection lives in computer-tool.node-resolution.test.ts.
- */
-import { createHash } from "node:crypto";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import type {
-  ComputerUseCapabilityDescriptor,
-  ComputerUseV2ActionName,
-} from "../../plugins/computer-use-contract.js";
-import type { AgentMessage } from "../runtime/index.js";
+import {
+  callGatewayToolMock,
+  COMPUTER_ACT_COMMAND,
+  type ComputerActBody,
+  type ComputerTool,
+  type ComputerToolOptions,
+  createVisionComputerTool,
+  EFFECTIVE_REF_WIDTH,
+  listNodesMock,
+  macComputerNode,
+  readFrameId,
+  readLastComputerActParams,
+  resetComputerToolMocks,
+  screenshotPayload,
+  TINY_PNG_BASE64,
+} from "./computer-tool.test-helpers.js";
 
-const listNodesMock = vi.fn();
-const callGatewayToolMock = vi.fn();
-const sleepMock = vi.hoisted(() => vi.fn());
-const TINY_PNG_BASE64 =
-  "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mP8/x8AAwMCAO+/p9sAAAAASUVORK5CYII=";
-const COMPUTER_ACT_COMMAND = "computer.act";
 const INVALID_SCROLL_AMOUNT = /scrollAmount must be a positive integer/;
 const INVALID_HOLD_DURATION = /duration must be >0 and <=10 seconds/;
 
-function imageIdentity(data: string, mimeType = "image/png") {
-  return createHash("sha256")
-    .update(JSON.stringify([mimeType, data]))
-    .digest("hex");
-}
-
-vi.mock("./nodes-utils.js", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("./nodes-utils.js")>();
-  return { ...actual, listNodes: listNodesMock };
-});
-
-vi.mock("./gateway.js", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("./gateway.js")>();
-  return { ...actual, callGatewayTool: callGatewayToolMock };
-});
-
-vi.mock("../../utils/sleep.js", () => ({ sleep: sleepMock }));
-
-const { createComputerTool, invalidateComputerFrameIfMissing } = await import("./computer-tool.js");
-const { DEFAULT_IMAGE_MAX_DIMENSION_PX } = await import("../image-sanitization.js");
-// With no config the reference width is capped at the default sanitization limit.
-const EFFECTIVE_REF_WIDTH = Math.min(1280, DEFAULT_IMAGE_MAX_DIMENSION_PX);
-
-function macComputerNode(overrides?: Record<string, unknown>) {
-  return {
-    nodeId: "mac-1",
-    displayName: "Studio",
-    platform: "macos",
-    connected: true,
-    commands: ["screen.snapshot", "computer.act"],
-    ...overrides,
-  };
-}
-
-function v2Descriptor(
-  actions: ComputerUseV2ActionName[],
-  overrides: Partial<ComputerUseCapabilityDescriptor> = {},
-): ComputerUseCapabilityDescriptor {
-  return {
-    contractVersion: 2 as const,
-    provider: { id: "fixture", label: "Fixture", generation: "generation-1" },
-    actions,
-    targets: ["screen", "window", "element", "browser"] as const,
-    deliveryModes: ["background", "foreground"] as const,
-    observations: ["image", "accessibility", "browser"] as const,
-    features: { recording: false, agentCursor: false, multiDisplay: false },
-    ...overrides,
-  };
-}
-
-function readActionEnum(tool: ComputerTool): string[] {
-  const schema = tool.parameters as { properties?: { action?: { enum?: string[] } } };
-  return schema.properties?.action?.enum ?? [];
-}
-
-function screenshotPayload(screenIndex = 0, base64 = TINY_PNG_BASE64) {
-  return {
-    payload: {
-      format: "png",
-      base64,
-      displayFrameId: `display-${screenIndex}-frame`,
-      width: 1280,
-      height: 800,
-      screenIndex,
-    },
-  };
-}
-
-function readFrameId(result: { details?: unknown }): string {
-  const frameId = (result.details as { frameId?: unknown } | undefined)?.frameId;
-  if (typeof frameId !== "string") {
-    throw new Error("missing frameId");
-  }
-  return frameId;
-}
-
-function readLastComputerActParams(): Record<string, unknown> {
-  const call = callGatewayToolMock.mock.calls.findLast(
-    (entry) => (entry[2] as { command?: string }).command === COMPUTER_ACT_COMMAND,
-  );
-  const body = call?.[2] as { params?: Record<string, unknown> } | undefined;
-  if (!body?.params) {
-    throw new Error("missing computer.act request");
-  }
-  const { executionId: _executionId, ...params } = body.params;
-  return params;
-}
-
 function expectedAct(action: string, fields: Record<string, unknown> = {}) {
   return { action, screenIndex: 0, refWidth: EFFECTIVE_REF_WIDTH, ...fields };
-}
-
-type ComputerTool = ReturnType<typeof createComputerTool>;
-type ComputerToolOptions = NonNullable<Parameters<typeof createComputerTool>[0]>;
-type ComputerActBody = {
-  nodeId?: string;
-  command?: string;
-  idempotencyKey?: string;
-  params?: Record<string, unknown>;
-};
-
-function createVisionComputerTool(options: ComputerToolOptions = {}) {
-  return createComputerTool({ modelHasVision: true, ...options });
 }
 
 function twoMacComputerNodes() {
@@ -142,7 +39,9 @@ function computerActBodies(): ComputerActBody[] {
         return body;
       }
       const { executionId: _executionId, ...params } = body.params;
-      return { ...body, params };
+      const sanitizedBody = Object.assign({}, body);
+      sanitizedBody.params = params;
+      return sanitizedBody;
     });
 }
 
@@ -243,474 +142,8 @@ async function executeComputerAction(params: Record<string, unknown>) {
   return readLastComputerActParams();
 }
 
-function computerToolResult(
-  toolCallId: string,
-  content: Extract<AgentMessage, { role: "toolResult" }>["content"],
-) {
-  return {
-    role: "toolResult" as const,
-    toolCallId,
-    toolName: "computer",
-    content,
-    details: {},
-    isError: false,
-    timestamp: 1,
-  } satisfies AgentMessage;
-}
-
-function trackedContextEpoch(value: number) {
-  return {
-    value,
-    frameToolCallId: "shot-1",
-    frameImageIdentity: imageIdentity(TINY_PNG_BASE64),
-  };
-}
-
-function screenshotToolResult(data = TINY_PNG_BASE64) {
-  return computerToolResult("shot-1", [{ type: "image", data, mimeType: "image/png" }]);
-}
-
-describe("computer screenshot context binding", () => {
-  it("keeps coordinates valid while the tracked tool result image remains visible", () => {
-    const contextEpoch = trackedContextEpoch(0);
-
-    expect(
-      invalidateComputerFrameIfMissing({
-        contextEpoch,
-        messages: [screenshotToolResult()],
-      }),
-    ).toBe(false);
-    expect(contextEpoch).toEqual(trackedContextEpoch(0));
-  });
-
-  it("expires coordinates once the final context drops the tracked image", () => {
-    const contextEpoch = trackedContextEpoch(0);
-
-    expect(
-      invalidateComputerFrameIfMissing({
-        contextEpoch,
-        messages: [computerToolResult("shot-1", [{ type: "text", text: "compacted" }])],
-      }),
-    ).toBe(true);
-    expect(contextEpoch).toEqual({ value: 1 });
-    expect(invalidateComputerFrameIfMissing({ contextEpoch, messages: [] })).toBe(false);
-    expect(contextEpoch.value).toBe(1);
-  });
-
-  it.each([
-    [
-      "expires coordinates when image input is disabled at the model boundary",
-      trackedContextEpoch(3),
-      [screenshotToolResult()],
-      true,
-      { value: 4 },
-    ],
-    [
-      "expires coordinates when middleware swaps the tracked screenshot",
-      trackedContextEpoch(5),
-      [screenshotToolResult("AQ==")],
-      undefined,
-      { value: 6 },
-    ],
-    [
-      "cleans up an orphaned image identity",
-      { value: 8, frameImageIdentity: imageIdentity(TINY_PNG_BASE64) },
-      [],
-      undefined,
-      { value: 9 },
-    ],
-  ])("%s", (_name, contextEpoch, messages, imagesBlocked, expected) => {
-    expect(invalidateComputerFrameIfMissing({ contextEpoch, messages, imagesBlocked })).toBe(true);
-    expect(contextEpoch).toEqual(expected);
-  });
-});
-
-describe("createComputerTool schema", () => {
-  it("keeps an undeclared node on the exact v1 action list", () => {
-    expect(readActionEnum(createComputerTool())).toEqual([
-      "screenshot",
-      "left_click",
-      "right_click",
-      "middle_click",
-      "double_click",
-      "triple_click",
-      "mouse_move",
-      "left_click_drag",
-      "left_mouse_down",
-      "left_mouse_up",
-      "scroll",
-      "type",
-      "key",
-      "hold_key",
-      "wait",
-    ]);
-  });
-
-  it("filters the model schema to a preselected v2 descriptor", () => {
-    const actions: ComputerUseV2ActionName[] = ["screenshot", "list_apps", "get_window_state"];
-    const tool = createComputerTool({ capabilityDescriptor: v2Descriptor(actions) });
-    expect(readActionEnum(tool)).toEqual(actions);
-  });
-
-  it("advertises resource actions only with an attempt cleanup owner", () => {
-    const actions: ComputerUseV2ActionName[] = ["browser_download", "start_recording"];
-    expect(
-      readActionEnum(createComputerTool({ capabilityDescriptor: v2Descriptor(actions) })),
-    ).toEqual([]);
-    expect(
-      readActionEnum(
-        createComputerTool({
-          capabilityDescriptor: v2Descriptor(actions),
-          registerRunCleanup: () => {},
-        }),
-      ),
-    ).toEqual(actions);
-  });
-
-  it("keeps the v2 guidance provider-neutral and free of host setup instructions", () => {
-    const description = createComputerTool({
-      capabilityDescriptor: v2Descriptor([
-        "screenshot",
-        "left_click",
-        "list_windows",
-        "get_window_state",
-        "set_value",
-      ]),
-    }).description;
-
-    expect(description).toContain("Observe first with `get_window_state`");
-    expect(description).toContain('`effect:"confirmed"` > `unverifiable` > `suspected_noop`');
-    expect(description).toContain("never blind-retry a mutation");
-    expect(description).toContain("untrusted input");
-    expect(description).not.toMatch(
-      /cua|peekaboo|\b(?:cli|mcp|daemon|socket|install(?:ation|ing)?)\b|verify_state|start_session|end_session|element_token|snapshot_id|window_id|delivery_mode/iu,
-    );
-    expect(description.length).toBeLessThan(2_400);
-  });
-
-  it("filters guidance to the selected node's advertised capability families", () => {
-    const desktopOnly = createComputerTool({
-      capabilityDescriptor: v2Descriptor(["screenshot", "left_click"], {
-        targets: ["screen"],
-        deliveryModes: ["foreground"],
-        observations: ["image"],
-      }),
-    }).description;
-    expect(desktopOnly).toContain("desktop coordinates from the latest screenshot");
-    expect(desktopOnly).toContain("stale frameId");
-    expect(desktopOnly).not.toMatch(
-      /get_window_state|accessibility|elementRef|window pixels|deliveryMode:"background"|background_unavailable/,
-    );
-
-    const windowBackground = createComputerTool({
-      capabilityDescriptor: v2Descriptor(
-        ["left_click", "list_windows", "get_window_state", "set_value"],
-        {
-          targets: ["window", "element"],
-          deliveryModes: ["background"],
-        },
-      ),
-    }).description;
-    expect(windowBackground).toContain(
-      "elementRef from the latest observation > window pixels from the latest window image",
-    );
-    expect(windowBackground).toContain('deliveryMode:"background"');
-    expect(windowBackground).toContain("background_occluded");
-    expect(windowBackground).not.toMatch(/desktop coordinates|foreground|frameId/);
-  });
-
-  it("publishes Codex-compatible fixed-size coordinate arrays", () => {
-    const properties = (
-      createComputerTool().parameters as {
-        properties?: Record<string, Record<string, unknown>>;
-      }
-    ).properties;
-
-    for (const key of ["coordinate", "startCoordinate"] as const) {
-      const schema = properties?.[key];
-      if (!schema) {
-        throw new Error(`missing ${key} schema`);
-      }
-      expect(schema).toMatchObject({
-        type: "array",
-        items: { type: "integer", minimum: 0 },
-        minItems: 2,
-        maxItems: 2,
-      });
-      expect(Array.isArray(schema.items)).toBe(false);
-      expect(schema).not.toHaveProperty("additionalItems");
-    }
-  });
-});
-
-describe("createComputerTool execution", () => {
-  beforeEach(() => {
-    listNodesMock.mockReset();
-    callGatewayToolMock.mockReset();
-    sleepMock.mockReset();
-    sleepMock.mockImplementation((ms: number, signal?: AbortSignal) => {
-      if (signal?.aborted) {
-        return Promise.reject(new Error("Aborted"));
-      }
-      if (ms === 500 || !signal) {
-        return Promise.resolve();
-      }
-      return new Promise<void>((_resolve, reject) => {
-        signal.addEventListener("abort", () => reject(new Error("Aborted")), { once: true });
-      });
-    });
-    listNodesMock.mockResolvedValue([macComputerNode()]);
-    callGatewayToolMock.mockImplementation(async (_method, _opts, body) =>
-      (body as ComputerActBody).command === COMPUTER_ACT_COMMAND
-        ? { payload: { ok: true } }
-        : screenshotPayload(),
-    );
-  });
-
-  it("rebuilds the visible action enum from the selected node declaration", async () => {
-    const actions: ComputerUseV2ActionName[] = ["screenshot", "list_apps", "get_window_state"];
-    listNodesMock.mockResolvedValue([macComputerNode({ computerUse: v2Descriptor(actions) })]);
-    const tool = createVisionComputerTool();
-    expect(readActionEnum(tool)).toHaveLength(15);
-    expect(tool.description).not.toContain("get_window_state");
-
-    await tool.execute("select", { action: "screenshot" });
-
-    expect(readActionEnum(tool)).toEqual(actions);
-    expect(tool.description).toContain("Observe first with `get_window_state`");
-  });
-
-  it("projects a provider observation without taking a duplicate desktop screenshot", async () => {
-    const actions: ComputerUseV2ActionName[] = ["get_window_state"];
-    listNodesMock.mockResolvedValue([macComputerNode({ computerUse: v2Descriptor(actions) })]);
-    callGatewayToolMock.mockResolvedValue({
-      payload: {
-        ok: true,
-        effect: "confirmed",
-        observation: {
-          kind: "window",
-          base64: TINY_PNG_BASE64,
-          format: "png",
-          width: 1,
-          height: 1,
-          observationId: "observation-1",
-          elements: [
-            {
-              elementRef: "element-1",
-              role: "button",
-              label: "Save",
-              bounds: { x: 0, y: 0, width: 1, height: 1 },
-            },
-          ],
-        },
-      },
-    });
-    const tool = createVisionComputerTool({ capabilityDescriptor: v2Descriptor(actions) });
-
-    const result = await tool.execute("observe", {
-      action: "get_window_state",
-      windowRef: "window-1",
-    });
-
-    expect(result.content).toContainEqual(
-      expect.objectContaining({ type: "image", mimeType: "image/png" }),
-    );
-    expect(callGatewayToolMock).toHaveBeenCalledOnce();
-    expect(readLastComputerActParams()).toEqual({
-      action: "get_window_state",
-      windowRef: "window-1",
-    });
-    expect(sleepMock).not.toHaveBeenCalledWith(500, expect.anything());
-  });
-
-  it("rejects stale semantic references before dispatch", async () => {
-    const actions: ComputerUseV2ActionName[] = ["get_window_state", "set_value"];
-    listNodesMock.mockResolvedValue([macComputerNode({ computerUse: v2Descriptor(actions) })]);
-    callGatewayToolMock.mockResolvedValue({
-      payload: {
-        ok: true,
-        observation: { kind: "window", observationId: "observation-current" },
-      },
-    });
-    const tool = createVisionComputerTool({ capabilityDescriptor: v2Descriptor(actions) });
-    await tool.execute("observe", { action: "get_window_state", windowRef: "window-1" });
-    callGatewayToolMock.mockClear();
-
-    await expect(
-      tool.execute("write", {
-        action: "set_value",
-        windowRef: "window-1",
-        elementRef: "element-1",
-        observationId: "observation-stale",
-        value: "hello",
-        deliveryMode: "background",
-      }),
-    ).rejects.toThrow("COMPUTER_STALE_OBSERVATION");
-    expect(callGatewayToolMock).not.toHaveBeenCalled();
-  });
-
-  it("maps browser observations and opaque refs through the public tool", async () => {
-    const actions: ComputerUseV2ActionName[] = ["get_browser_state", "browser_pointer"];
-    listNodesMock.mockResolvedValue([macComputerNode({ computerUse: v2Descriptor(actions) })]);
-    callGatewayToolMock.mockResolvedValueOnce({
-      payload: {
-        ok: true,
-        observation: { kind: "browser", observationId: "browser-observation-1" },
-        details: {
-          browserRef: "browser-1",
-          pageRef: "page-1",
-          elements: [{ elementRef: "element-1" }, { elementRef: "element-2" }],
-        },
-      },
-    });
-    const tool = createVisionComputerTool({ capabilityDescriptor: v2Descriptor(actions) });
-
-    await tool.execute("observe-browser", {
-      action: "get_browser_state",
-      browserRef: "browser-1",
-      pageRef: "page-1",
-      snapshotFormat: "dom_refs_v1",
-      includeScreenshot: true,
-    });
-    expect(readLastComputerActParams()).toEqual({
-      action: "get_browser_state",
-      browserRef: "browser-1",
-      pageRef: "page-1",
-      snapshotFormat: "dom_refs_v1",
-      includeScreenshot: true,
-    });
-
-    callGatewayToolMock.mockImplementation(async (_method, _opts, body) =>
-      (body as ComputerActBody).command === COMPUTER_ACT_COMMAND
-        ? { payload: { ok: true, effect: "confirmed" } }
-        : screenshotPayload(),
-    );
-    await tool.execute("drag-browser", {
-      action: "browser_pointer",
-      browserRef: "browser-1",
-      pageRef: "page-1",
-      observationId: "browser-observation-1",
-      pointerAction: "drag",
-      inputRoute: "dom_event",
-      elementRef: "element-1",
-      destinationElementRef: "element-2",
-    });
-    expect(readLastComputerActParams()).toEqual({
-      action: "browser_pointer",
-      browserRef: "browser-1",
-      pageRef: "page-1",
-      observationId: "browser-observation-1",
-      pointerAction: "drag",
-      inputRoute: "dom_event",
-      elementRef: "element-1",
-      destinationElementRef: "element-2",
-    });
-  });
-
-  it("routes an observation-bound element click without requiring coordinates", async () => {
-    const actions: ComputerUseV2ActionName[] = ["get_window_state", "left_click"];
-    listNodesMock.mockResolvedValue([macComputerNode({ computerUse: v2Descriptor(actions) })]);
-    callGatewayToolMock.mockImplementation(async (_method, _opts, body) => {
-      const request = body as ComputerActBody;
-      if (request.command !== COMPUTER_ACT_COMMAND) {
-        return screenshotPayload();
-      }
-      if (request.params?.action === "get_window_state") {
-        return {
-          payload: {
-            ok: true,
-            observation: {
-              kind: "window",
-              observationId: "observation-1",
-            },
-          },
-        };
-      }
-      return { payload: { ok: true, effect: "confirmed" } };
-    });
-    const tool = createVisionComputerTool({ capabilityDescriptor: v2Descriptor(actions) });
-    await tool.execute("observe", { action: "get_window_state", windowRef: "window-1" });
-
-    await expect(
-      tool.execute("click", {
-        action: "left_click",
-        windowRef: "window-1",
-        elementRef: "element-1",
-        observationId: "observation-1",
-        deliveryMode: "background",
-      }),
-    ).resolves.toBeDefined();
-    expect(readLastComputerActParams()).toEqual({
-      action: "left_click",
-      screenIndex: 0,
-      refWidth: EFFECTIVE_REF_WIDTH,
-      windowRef: "window-1",
-      elementRef: "element-1",
-      observationId: "observation-1",
-      deliveryMode: "background",
-    });
-  });
-
-  it("maps the recording family through opaque resource parameters", async () => {
-    const actions: ComputerUseV2ActionName[] = [
-      "get_recording_state",
-      "start_recording",
-      "stop_recording",
-      "replay_trajectory",
-    ];
-    listNodesMock.mockResolvedValue([macComputerNode({ computerUse: v2Descriptor(actions) })]);
-    const tool = createVisionComputerTool({
-      capabilityDescriptor: v2Descriptor(actions),
-      registerRunCleanup: () => {},
-    });
-    const resourceHandle = "openclaw:computer-resource:v1:123e4567-e89b-42d3-a456-426614174000";
-
-    await tool.execute("record", { action: "start_recording", recordVideo: true });
-    expect(readLastComputerActParams()).toEqual({ action: "start_recording", recordVideo: true });
-    await tool.execute("replay", {
-      action: "replay_trajectory",
-      resourceHandle,
-      delayMs: 25,
-      stopOnError: false,
-    });
-    expect(readLastComputerActParams()).toEqual({
-      action: "replay_trajectory",
-      resourceHandle,
-      delayMs: 25,
-      stopOnError: false,
-    });
-  });
-
-  it("closes the exact host execution through attempt-owned cleanup", async () => {
-    const actions: ComputerUseV2ActionName[] = ["start_recording"];
-    listNodesMock.mockResolvedValue([macComputerNode({ computerUse: v2Descriptor(actions) })]);
-    let cleanup: ((reason: string) => Promise<void>) | undefined;
-    const tool = createVisionComputerTool({
-      capabilityDescriptor: v2Descriptor(actions),
-      registerRunCleanup: (registered) => {
-        cleanup = registered;
-      },
-    });
-
-    await tool.execute("record", { action: "start_recording" });
-    const start = callGatewayToolMock.mock.calls
-      .map((call) => call[2] as ComputerActBody)
-      .findLast((body) => body.command === COMPUTER_ACT_COMMAND);
-    if (!start?.params) {
-      throw new Error("missing start_recording node invocation");
-    }
-    const executionId = start.params.executionId;
-    expect(executionId).toEqual(expect.any(String));
-
-    await cleanup?.("completion");
-
-    const close = callGatewayToolMock.mock.calls.at(-1)?.[2] as ComputerActBody;
-    expect(close.params).toEqual({
-      action: "__close_execution",
-      executionId,
-      reason: "completion",
-    });
-  });
+describe("createComputerTool v1 execution", () => {
+  beforeEach(resetComputerToolMocks);
 
   it.each([
     [
